+++ libftl/media.c
@@ -85,8 +85,7 @@ ftl_status_t media_init(ftl_stream_configuration_private_t *ftl) {
 		clear_stats(&comp->stats);
 	}
 
-	//ftl->video.media_component.timestamp_step = (uint32_t)(90000.f / ftl->video.frame_rate);
-	ftl->video.media_component.timestamp_step = (uint32_t)(90000.f * ftl->video.fps_den / ftl->video.fps_num );
+	ftl->video.media_component.timestamp_step = 90000.f * (float)ftl->video.fps_den / (float)ftl->video.fps_num;
 	ftl->video.wait_for_idr_frame = TRUE;
 	ftl->audio.media_component.timestamp_step = 48000 / 50; //TODO: dont assume the step size for audio
 
@@ -556,7 +555,7 @@ static int _media_make_video_rtp_packet(ftl_stream_configuration_private_t *ftl,
 
 	rtp_header = htonl((2 << 30) | (mc->payload_type << 16) | mc->seq_num);
 	*((uint32_t*)out)++ = rtp_header;
-	rtp_header = htonl(mc->timestamp);
+	rtp_header = htonl((uint32_t)mc->timestamp);
 	*((uint32_t*)out)++ = rtp_header;
 	rtp_header = htonl(mc->ssrc);
 	*((uint32_t*)out)++ = rtp_header;
@@ -606,7 +605,7 @@ static int _media_make_audio_rtp_packet(ftl_stream_configuration_private_t *ftl,
 
 	rtp_header = htonl((2 << 30) | (1 << 23) | (mc->payload_type << 16) | mc->seq_num);
 	*((uint32_t*)out)++ = rtp_header;
-	rtp_header = htonl(mc->timestamp);
+	rtp_header = htonl((uint32_t)mc->timestamp);
 	*((uint32_t*)out)++ = rtp_header;
 	rtp_header = htonl(mc->ssrc);
 	*((uint32_t*)out)++ = rtp_header;
